+++ cli/main_test.go
@@ -243,8 +243,8 @@ func newTestRuntimeConfig(dir, consolePath string, create bool) (oci.RuntimeConf
 		HypervisorType:   vc.QemuHypervisor,
 		HypervisorConfig: hypervisorConfig,
 		AgentType:        vc.KataContainersAgent,
-		ProxyType:        vc.CCProxyType,
-		ShimType:         vc.CCShimType,
+		ProxyType:        vc.KataProxyType,
+		ShimType:         vc.KataShimType,
 		Console:          consolePath,
 	}, nil
 }
