+++ pkg/katautils/create_test.go
@@ -121,8 +121,8 @@ func newTestRuntimeConfig(dir, consolePath string, create bool) (oci.RuntimeConf
 		HypervisorType:   vc.QemuHypervisor,
 		HypervisorConfig: hypervisorConfig,
 		AgentType:        vc.KataContainersAgent,
-		ProxyType:        vc.CCProxyType,
-		ShimType:         vc.CCShimType,
+		ProxyType:        vc.KataProxyType,
+		ShimType:         vc.KataShimType,
 		Console:          consolePath,
 	}, nil
 }
