+++ virtcontainers/kata_shim_test.go
@@ -12,14 +12,23 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"syscall"
 	"testing"
 	"time"
+	"unsafe"
 
 	. "github.com/kata-containers/runtime/virtcontainers/pkg/mock"
 )
 
-var testKataShimPath = "/usr/bin/virtcontainers/bin/test/kata-shim"
-var testKataShimProxyURL = "foo:///foo/kata-containers/proxy.sock"
+const (
+	testContainer = "testContainer"
+)
+
+var (
+	testKataShimPath     = "/usr/bin/virtcontainers/bin/test/kata-shim"
+	testKataShimProxyURL = "foo:///foo/kata-containers/proxy.sock"
+	testWrongConsolePath = "/foo/wrong-console"
+)
 
 func getMockKataShimBinPath() string {
 	if DefaultMockKataShimBinPath == "" {
@@ -270,6 +279,54 @@ func TestKataShimStartWithConsoleNonExistingFailure(t *testing.T) {
 	}
 
 	testKataShimStart(t, sandbox, params, true)
+}
+
+func ioctl(fd uintptr, flag, data uintptr) error {
+	if _, _, err := syscall.Syscall(syscall.SYS_IOCTL, fd, flag, data); err != 0 {
+		return err
+	}
+
+	return nil
+}
+
+// unlockpt unlocks the slave pseudoterminal device corresponding to the master pseudoterminal referred to by f.
+func unlockpt(f *os.File) error {
+	var u int32
+
+	return ioctl(f.Fd(), syscall.TIOCSPTLCK, uintptr(unsafe.Pointer(&u)))
+}
+
+// ptsname retrieves the name of the first available pts for the given master.
+func ptsname(f *os.File) (string, error) {
+	var n int32
+
+	if err := ioctl(f.Fd(), syscall.TIOCGPTN, uintptr(unsafe.Pointer(&n))); err != nil {
+		return "", err
+	}
+
+	return fmt.Sprintf("/dev/pts/%d", n), nil
+}
+
+func newConsole() (*os.File, string, error) {
+	master, err := os.OpenFile("/dev/ptmx", syscall.O_RDWR|syscall.O_NOCTTY|syscall.O_CLOEXEC, 0)
+	if err != nil {
+		return nil, "", err
+	}
+
+	console, err := ptsname(master)
+	if err != nil {
+		return nil, "", err
+	}
+
+	if err := unlockpt(master); err != nil {
+		return nil, "", err
+	}
+
+	if err := os.Chmod(console, 0600); err != nil {
+		return nil, "", err
+	}
+
+	return master, console, nil
 }
 
 func TestKataShimStartWithConsoleSuccessful(t *testing.T) {
