+++ virtcontainers/pkg/oci/utils_test.go
@@ -107,9 +107,9 @@ func TestMinimalSandboxConfig(t *testing.T) {
 
 	runtimeConfig := RuntimeConfig{
 		HypervisorType: vc.QemuHypervisor,
-		AgentType:      vc.HyperstartAgent,
-		ProxyType:      vc.CCProxyType,
-		ShimType:       vc.CCShimType,
+		AgentType:      vc.KataContainersAgent,
+		ProxyType:      vc.KataProxyType,
+		ShimType:       vc.KataShimType,
 		Console:        consolePath,
 	}
 
@@ -222,9 +222,9 @@ func TestMinimalSandboxConfig(t *testing.T) {
 		Hostname: "testHostname",
 
 		HypervisorType: vc.QemuHypervisor,
-		AgentType:      vc.HyperstartAgent,
-		ProxyType:      vc.CCProxyType,
-		ShimType:       vc.CCShimType,
+		AgentType:      vc.KataContainersAgent,
+		ProxyType:      vc.KataProxyType,
+		ShimType:       vc.KataShimType,
 
 		NetworkConfig: expectedNetworkConfig,
